src/requirements.txt
cassandra-driver==3.17.0
mysqlclient==1.4.2.post1
pika==1.0.0
redis==3.2.1
